Extract process and object stuff off profiler
In the future more things other stuff such as process discovery,
executable mapping logic, etc could be extracted too

Test Plan
=========

CI
