+++ src/lib.rs
@@ -2,6 +2,7 @@ pub mod bpf;
 pub mod collector;
 pub mod ksym;
 pub mod perf_events;
+pub mod process;
 pub mod profile;
 pub mod profiler;
 pub mod unwind_info;
