+++ src/profile.rs
@@ -13,12 +13,12 @@ use tracing::{debug, error, span, Level};
 
 use crate::ksym::Ksym;
 use crate::ksym::KsymIter;
+use crate::process::ObjectFileInfo;
+use crate::process::ProcessInfo;
 use crate::profiler::AggregatedProfile;
 use crate::profiler::AggregatedSample;
 use crate::profiler::Frame;
 use crate::profiler::FrameAddress;
-use crate::profiler::ObjectFileInfo;
-use crate::profiler::ProcessInfo;
 use crate::profiler::RawAggregatedProfile;
 use crate::usym::symbolize_native_stack_blaze;
 use lightswitch_object::ExecutableId;
